alteração do filtro para o comp. ItemListContainer
