+++ src/Router.jsx
@@ -7,6 +7,7 @@ import { Admin } from "./pages/admin";
 import { collection, getDocs, getFirestore } from "firebase/firestore";
 import app from "./services/firebase";
 import { useEffect, useState } from "react";
+import { Vitrine } from "./pages/Vitrine";
 
 export function Router() {
 
@@ -77,7 +78,7 @@ export function Router() {
             element={
                 <DefaultLayout />
             } >
-                <Route path="/" element={ <ItemListContainer items={productList} /> } />
+                <Route path="/" element={ <Vitrine /> } />
                 <Route path="/category/:categoryId" element={ <ItemListContainer items={productList} /> } />
                 <Route path="/item/:itemId" element={ <ItemDetailContainer items={productList} /> } />
                 <Route path="/cart" element={ <Cart /> } />
