+++ src/components/ItemListContainer/index.jsx
@@ -1,4 +1,5 @@
 
+import { useEffect, useState } from "react";
 import { ItemList } from "../ItemList";
 import style from "./style.module.css";
 import { useParams } from "react-router-dom";
@@ -7,11 +8,41 @@ import { useParams } from "react-router-dom";
 export function ItemListContainer({items}) {
     const { categoryId } = useParams();
 
+    const [produtos, setProdutos] = useState();
+  const [loading, setLoading] = useState(true);
+
+
+  function filtroProdutos() {
+    if(categoryId) {
+      console.log("tentando filtrar...")
+      var itensFiltrados = items.filter(function(i) {
+        console.log(i);
+        return i.categoryId == categoryId;
+      })
+      setProdutos(itensFiltrados);
+      setLoading(false);
+    } else {
+      setProdutos(items);
+      setLoading(false);
+    }
+  }
+
+    useEffect(() => {
+        console.log(categoryId);
+        console.log(loading);
+        //console log OK - entrando
+        filtroProdutos();
+    },[loading, categoryId])
+
     return (
         <>
             <div className={style.container}></div>
             <br/>
-            <ItemList categoryId={categoryId} items={items}/>
+            {!loading ? 
+                <ItemList categoryId={categoryId} items={produtos}/> :
+                "Carregando..."
+            }
+            
         </>
     );
 }
